Moved app.py to horizon module.
