+++ dipy/viz/tests/test_apps.py
@@ -16,7 +16,7 @@ fury, has_fury, setup_module = optional_package('fury')
 if has_fury:
     from fury import window
 
-    from dipy.viz.app import horizon
+    from dipy.viz.horizon.app import horizon
 
 skip_it = use_xvfb == 'skip'
 
@@ -114,7 +114,7 @@ def test_horizon():
                 clusters_lt=np.inf, clusters_gt=0,
                 world_coords=False, interactive=False)
 
-    msg = 'Currently native coordinates are not supported for streamlines'
+    msg = 'Currently native coordinates are not supported for streamlines.'
     npt.assert_(msg in str(ve.exception))
 
     # only images
@@ -140,6 +140,7 @@ def test_roi_images():
     img3 = np.zeros((5, 5, 5))
     img3[3, 3, 3] = 1
     images = [(img1, np.eye(4)), (img2, np.eye(4)), (img3, np.eye(4))]
+    # TODO: Fix this. Fails due to multi-image feature
     show_m = horizon(images=images, return_showm=True)
     analysis = window.analyze_scene(show_m.scene)
     npt.assert_equal(analysis.actors, 0)
