+++ dipy/workflows/viz.py
@@ -2,7 +2,7 @@ import numpy as np
 from os.path import join as pjoin
 from dipy.workflows.workflow import Workflow
 from dipy.io.image import load_nifti
-from dipy.viz.app import horizon
+from dipy.viz.horizon.app import horizon
 from dipy.io.peaks import load_peaks
 from dipy.io.streamline import load_tractogram
 from dipy.io.utils import create_nifti_header
